StdLib/Include/sys/syslimits.h: Increase OPEN_MAX from 20 to 255
REF: https://bugzilla.tianocore.org/show_bug.cgi?id=3548

A port of Python 3.6.8 interpreter for UEFI uses the
standard C Library from the edk2-libc project. The python
interpreter fails with the following error when the
python interpreter is launched:

"IOError: [Errno 24] Too many open files"

Python 3.6.8 interpreter opens and loads many python modules
(files) when it is boot strapped and this is leading to the
too many open files error.

To fix the issue the OPEN_MAX limit value is increased from
20 to 255.

Cc: Daryl McDaniel <edk2-lists@mc2research.org>
Cc: Jaben Carsey <jaben.carsey@intel.com>
Cc: Michael D Kinney <michael.d.kinney@intel.com>
Cc: Rebecca Cran <rebecca@nuviainc.com>
Signed-off-by: Jayaprakash Nevara <n.jayaprakash@intel.com>