StdLib/LibC/gdtoa: Remove ldtoa() function that is not used.
Removing the ldtoa.c file from the INF removes the ldtoa()
function from this library.  This function is not used and
when /WHOLEARCHIVE linker flag is enabled this unused
function makes references to other functions that are not
implemented and that generated a link failure.  Removing the
source file resolves the link failure.

Signed-off-by: Michael D Kinney <michael.d.kinney@intel.com>